rename: Make directory for each calendar
