+++ components/MonthCalendar/MonthCalendar.module.scss
@@ -1,4 +1,4 @@
-@use '../styles/variables.scss' as v;
+@use '../../styles/variables.scss' as v;
 
 .wrapper {
     display: flex;
