+++ pages/day/[year]/[month]/[date].tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import React, { useEffect } from 'react';
 
-import DayCalendar from '@components/DayCalendar';
+import DayCalendar from '@components/DayCalendar/DayCalendar';
 import { CalendarType, useDateContext } from '@contexts/DateContext';
 
 export default function DayPage() {
