+++ pages/month/today.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 
-import MonthCalendar from '@components/MonthCalendar';
+import MonthCalendar from '@components/MonthCalendar/MonthCalendar';
 import { CalendarType, useDateContext } from '@contexts/DateContext';
 
 export default function MonthTodayPage() {
