+++ pages/schedule/today.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 
-import ScheduleCalendar from '@components/ScheduleCalendar';
+import ScheduleCalendar from '@components/ScheduleCalendar/ScheduleCalendar';
 import { CalendarType, useDateContext } from '@contexts/DateContext';
 
 export default function ScheduleTodayPage() {
