+++ pages/week/[year]/[month]/[date].tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import React, { useEffect } from 'react';
 
-import WeekCalendar from '@components/WeekCalendar';
+import WeekCalendar from '@components/WeekCalendar/WeekCalendar';
 import { CalendarType, useDateContext } from '@contexts/DateContext';
 
 export default function WeekPage() {
